+++ Source/SelfService/Web/integrations/connection/connectionDetails/commands/commandsList/commandsListView/commandsListDataGrid/CommandsListDetailPanel.tsx
